+++ Order.Management/Shapes/Triangle.cs
@@ -4,7 +4,7 @@
     {
         public override ShapeName Name => ShapeName.Triangle;
 
-        public override int Price => 2;
+        public override decimal Price => 2;
 
         public Triangle(int numberOfRedTriangles, int numberOfBlueTriangles, int numberOfYellowTriangles) : base(numberOfRedTriangles, numberOfBlueTriangles, numberOfYellowTriangles) { }
     }
